% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqdom.fda-package.R
\docType{package}
\name{freqdom.fda-package}
\alias{freqdom.fda-package}
\alias{_PACKAGE}
\alias{fda.ts}
\title{Functional time series: dynamic FPCA}
\description{
Implementation of dynamic functional principle component analysis (FDPCA),
simulation of FAR and FMA processes and frequency domain tools for funcional data.
The package is a wrapper for functionality of the multivariate package \pkg{freqdom}
for applying frequency domain on objects from \pkg{fda}.
Compared to \pkg{freqdom} multiple new visualization methods are added --
adequate only if data has functional structure.
}
\details{
\pkg{fda.ts} package allows you to functional time series objects
in both time and frequency domains. We implement dynamic functional principal component analysis
methods, enabling spectral decomposition of a stationary functional time series into uncorrelated components.

Dynamic functional principal component analysis enables estimation of temporal filters which
transform a functional time series into a vector time series with uncorrelated components,
maximizing the long run variance explained.
There are two key differnces between classical PCA and dynamic PCA:
\itemize{
\item Components returned by the dynamic procedure are uncorrelated in time, i.e. for any \eqn{i \neq j}
and \eqn{l \in Z}, \eqn{Y_i(t)} and \eqn{Y_j(t_l)} are uncorrelated,
\item The mapping maximizes the long run variance, which, in case of stationary functional time series, means
that the process reconstructed from and \eqn{d > 0} first dynamic principal components
better approximates your functional time series process than the first \eqn{d} classic principal components.
}

Thanks to the unique structure of functional data, where multivariate components are on the
same scale and are directly related to each other, one can visualize properties of the
filters and frequency domain tools by presenting 2D kernels.

For details, please refer to literature below and to help pages of functions \code{\link{fts.dpca}}
for estimating the components, \code{\link{fts.dpca.scores}} for estimating scores and
\code{\link{fts.dpca.KLexpansion}} for retrieving the signal from components.

Apart from frequency domain techniques for stationary functional time series,
\pkg{freqdom} provides a toolset of operators such as a functional spectral density operator
(\code{\link{fts.spectral.density}}) as well as simulation of functional time series
models \code{\link{fts.rar}}, \code{\link{fts.rma}} generating functional
autoregressive and moving average respectively.
These functions enable developing new techniques based on the Frequency domain analysis.
}
\references{
Hormann Siegfried, Kidzinski Lukasz and Hallin Marc.
\emph{Dynamic functional principal components.} Journal of the Royal
Statistical Society: Series B (Statistical Methodology) 77.2 (2015): 319-348.

Hormann Siegfried, Kidzinski Lukasz and Kokoszka Piotr.
\emph{Estimation in functional lagged regression.}
Journal of Time Series Analysis 36.4 (2015): 541-561.

Hormann Siegfried and Kidzinski Lukasz.
\emph{A note on estimation in Hilbertian linear models.}
Scandinavian journal of statistics 42.1 (2015): 43-62.
}
