% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fts.dpca.KLexpansion.R
\name{fts.dpca.KLexpansion}
\alias{fts.dpca.KLexpansion}
\title{Dynamic KL expansion}
\usage{
fts.dpca.KLexpansion(X, dpcs = fts.dpca.filters(fts.spectral.density(X)))
}
\arguments{
\item{X}{a functional time series given as an object of class \code{\link[fda]{fd}}.}

\item{dpcs}{an object of class \code{fts.timedom}, representing the dpca filters
obtained from the sample \code{X}. If \code{dpsc = NULL}, then \code{dpcs = fts.dpca.filter(fts.spectral.density(X))} is used.}
}
\value{
An object of class \code{\link[fda]{fd}}.
}
\description{
Computes the dynamic KL expansion up to a given order.
}
\details{
This function computes the \eqn{L}-order dynamic functional principal components expansion, defined by
\deqn{
  \hat{X}_{t}^L(u):=\sum_{\ell=1}^L\sum_{k\in\mathbf{Z}}  Y_{\ell,t+k} \phi_{\ell k}(u),\quad 1\leq L\leq d,
}
where \eqn{\phi_{\ell k}(v)} and \eqn{d} are explained in  \code{fts.dpca.filters} and \eqn{Y_{\ell k}} are the dynamic functional PC scores as in \code{fts.dpca.scores}. For the sample version the sum extends over the range of lags for which the \eqn{\phi_{\ell k}} are defined.

For more details we refer to Hormann et al. (2015).
}
\references{
Hormann, S., Kidzinski, L., and Hallin, M.
\emph{Dynamic functional principal components.} Journal of the Royal
Statistical Society: Series B (Statistical Methodology) 77.2 (2015): 319-348.
}
\seealso{
The multivariate equivalent in the \code{freqdom} package: \code{\link[freqdom]{dpca.KLexpansion}}
}
\keyword{DPCA}
