% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fts.plot.operators.lukasz.R
\name{fts.plot.operators.lukasz}
\alias{fts.plot.operators.lukasz}
\title{Plot kernels}
\usage{
fts.plot.operators.lukasz(X, cor = FALSE, grid = FALSE,
  main = "Operators")
}
\description{
Plot kernels
}
