% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.fts.freqdom.R
\name{is.fts.freqdom}
\alias{is.fts.freqdom}
\title{Checks if an object belongs to the class fts.freqdom}
\usage{
is.fts.freqdom(X)
}
\arguments{
\item{X}{some object}
}
\value{
\code{TRUE} if \code{X} is of type \code{\link{fts.freqdom}}, \code{FALSE} otherwise
}
\description{
Checks if an object belongs to the class \code{\link{fts.freqdom}}.
}
\seealso{
\code{\link{fts.freqdom}}, \code{\link{fts.timedom}}, \code{\link{is.fts.timedom}}
}
\keyword{classes}
