% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.fts.timedom.R
\name{is.fts.timedom}
\alias{is.fts.timedom}
\title{Checks if an object belongs to the class fts.timedom}
\usage{
is.fts.timedom(X)
}
\arguments{
\item{X}{some object}
}
\value{
\code{TRUE} if \code{X} is of type \code{\link{fts.timedom}}, \code{FALSE} otherwise
}
\description{
Checks if an object belongs to the class \code{\link{fts.timedom}}.
}
\seealso{
\code{\link{fts.freqdom}}, \code{\link{fts.timedom}}, \code{\link{is.fts.freqdom}}
}
\keyword{classes}
