% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm10.R
\docType{data}
\name{pm10}
\alias{pm10}
\title{PM10 dataset}
\format{\itemize{
\item 
}

\describe{
\item{\code{Gyro. X}}{Gyroscope X-Axis}
\item{\code{Gyro. Y}}{Gyroscope Y-Axis}
\item{\code{Gyro. Z}}{Gyroscope Z-Axis}
\item{\code{Accel. X}}{Accelerometer X-Axis}
\item{\code{Accel. Y}}{Accelerometer Y-Axis}
\item{\code{Accel. Z}}{Accelerometer Z-Axis}
}}
\source{
Graz
}
\usage{
pm10
}
\description{
PM10 dataset
}
\examples{
\dontrun{
data(pm10)
summary(pm10)
}
}
\author{
JJB
}
\keyword{datasets}
