% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fts.plot.covariance.R
\name{fts.plot.covariance}
\alias{fts.plot.covariance}
\title{Plot covariance and cross covariance kernels of functional time series}
\usage{
fts.plot.covariance(X, Y = X, cor = FALSE, main = "Operators",
  res = 200, lags = -1:1)
}
\description{
Plot covariance and cross covariance kernels of functional time series
}
